+++ alinkjs/packages/assetlink-default-plugins/plugins/AssetPage.alink.vue
@@ -21,8 +21,12 @@ watch(metaActionDefs, () => {
 
 <template alink-route[net.symbioquine.farmos_asset_link.routes.v0.asset_page]="/asset/:assetRef">
   <q-page padding class="text-left">
-    <asset-resolver :asset-ref="$route.params.assetRef" #default="{ asset }" @asset-resolved="resolvedAsset = $event">
-        <h4 class="q-my-xs">Asset: <render-widget
+    <asset-resolver
+      :asset-ref="$route.params.assetRef"
+      #default="{ asset }"
+      @asset-resolved="resolvedAsset = $event"
+      :key="$route.params.assetRef">
+        <h4 class="q-my-xs"><render-widget
               name="asset-name"
               :context="{ asset }"
               >{{ asset.attributes.name }}</render-widget>